Player/Enemy flip bug fixes
Fixed Player not flipping when DodgeRolling right after ShieldBashing.
Flip would be disabled, but movement allowed player to have full control of the dodge roll.

Increased speed of DodgeRoll and lowered the time to complete the roll.

----------------------------
Added a separate raycast length of back facing Enemy raycast.
This prevents enemy from flipping repeatedly when the player is on the other side of a wall.
*This is a temporary fix, and won't work with small platforms. Raycast will detect player, turn around because of WallDetect, then detect the player with the back facing raycast and repeat...