Stage Clear added
Reduced freezeTime on hits.

Re-added XP, added gold to player on enemy kills.
Added a StageClearCheck object to count enemies in the stage and allow the player to move to the next stage on clearing.

Created a Tutorial Stage.
The EndPortal script will have a 50/50 to choose between 2 stages to send the player to.