new Scenes added
New scenes added as stages.
LevelLoader moves player object to each scene.
StageClear counts all the enemies in the scene, once cleared it will allow the player to continue to the next stage.

LevelLoader is still in-progress, current references like TimeManager, HitEffects, etc are broken when moving to the next stage.