Finished MainMenu load
When starting the game from the MainMenu scene, the Player scene (_NeverUnload) is loaded first.
Once the Player scene is loaded, it is set to the active scene. MainMenu is then unloaded.
TutorialStage is then loaded.

Added a small collider platform under the Player in the Player Scene. Sometimes the TutorialStage wouldn't load fast enough and the Player would fall.